requirements.txt
selenium==4.16.0
beautifulsoup4==4.12.2
lxml==5.1.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
nltk==3.8.1
wordcloud==1.9.3
scikit-learn>=1.0.0
google-generativeai==0.3.2
httpx
parsel
loguru
